Starting to pull out debugger, still a little broken.
